iox-#378 initial RouDi startup test
