+++ .gitignore
@@ -12,3 +12,4 @@ clangd/
 site/
 doc/website/API-reference/
 *.project
+__pycache__
